pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snowflake-mcp-server"
version = "1.0.0"
description = "MCP Server for Snowflake database integration with Kiro CLI"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.7.1",
    "python-dotenv>=1.0.0",
    "snowflake-connector-python>=3.15.0",
    "flask>=3.0.0",
    "plotly>=5.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.25.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["logs*", "tests*"]

[project.scripts]
snowflake-mcp-server = "src.main:main"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=html --cov-report=term-missing"
